+++ shared/src/commonMain/kotlin/dev/tekofx/pinchodownloader/ui/components/AppTitle.kt
@@ -1,9 +1,11 @@
 package dev.tekofx.pinchodownloader.ui.components
 
+import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.onClick
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -14,8 +16,10 @@ import org.jetbrains.compose.resources.painterResource
 import pinchodownloader.shared.generated.resources.Res
 import pinchodownloader.shared.generated.resources.icon
 
+@OptIn(ExperimentalFoundationApi::class)
 @Composable
 fun AppTitle(
+    onIconClick: () -> Unit,
 ) {
     Row(
         verticalAlignment = Alignment.Bottom,
@@ -23,7 +27,7 @@ fun AppTitle(
     ) {
         Image(
             painter = painterResource(Res.drawable.icon), contentDescription = null,
-            Modifier.size(40.dp)
+            Modifier.size(40.dp).onClick(onClick = onIconClick)
         )
         Text("Pincho Downloader", style = MaterialTheme.typography.displaySmall)
         UpdateTag()