+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ### Added
 
 - Format tag in video details
+- Added log view
+- Make video text trim to not take all width